requirements.txt
asgiref==3.8.1
astroid==3.3.8
aws_secretsmanager_caching==1.1.3
boto3==1.37.23
botocore==1.37.23
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
Django==5.1.5
django-environ==0.12.0
django-filter==24.3
django-storages==1.14.5
djangorestframework==3.15.2
drf-yasg==1.21.8
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isort==6.0.0
jmespath==1.0.1
mccabe==0.7.0
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.4
pylint-plugin-utils==0.8.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
setuptools==75.8.0
setuptools-scm==8.2.0
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
